services/auth/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
